Apply suggestions from code review

+++ wled00/um_manager.cpp
@@ -68,7 +68,7 @@ bool UsermodManager::onEspNowMessage(uint8_t* sender, uint8_t* payload, uint8_t 
   return false;
 }
 #endif
-bool UsermodManager::onUdpPacket(uint8_t* payload, uint8_t len) {
+bool UsermodManager::onUdpPacket(uint8_t* payload, size_t len) {
   for (auto mod = _usermod_table_begin; mod < _usermod_table_end; ++mod) if ((*mod)->onUdpPacket(payload, len)) return true;
   return false;
 }
